@@ -12,6 +12,9 @@
   },
   "private": false,
   "dependencies": {
+    "ng-click-outside": "^4.0.0"
+  },
+  "peerDependencies": {
     "@angular/common": "^6.1.4",
     "@angular/compiler": "^6.1.4",
     "@angular/core": "^6.1.4",
@@ -19,10 +22,7 @@
     "@angular/platform-browser": "^6.1.4",
     "@angular/platform-browser-dynamic": "^6.1.4",
     "core-js": "^2.5.7",
-    "ng-click-outside": "^4.0.0",
     "npm-audit": "0.0.3",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
@@ -44,7 +44,7 @@
     "ng-packagr": "^4.1.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.11.0",
-    "typescript": "~2.7.0"
+    "typescript": "~2.9.2"
   },
   "publishConfig": {
     "tag": "y-generic-list"
