@@ -11,57 +11,57 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/animations": "^6.1.4",
+    "@angular/common": "^6.1.4",
+    "@angular/compiler": "^6.1.4",
+    "@angular/core": "^6.1.4",
+    "@angular/forms": "^6.1.4",
+    "@angular/http": "^6.1.4",
+    "@angular/platform-browser": "^6.1.4",
+    "@angular/platform-browser-dynamic": "^6.1.4",
+    "@angular/router": "^6.1.4",
     "@kyma-project/y-generic-list": "^0.4.10",
-    "@types/async": "^2.0.45",
-    "angular-oauth2-oidc": "^3.1",
-    "angular-sortablejs": "^2.1.0",
-    "core-js": "^2.4.1",
+    "@types/async": "^2.0.49",
+    "rxjs": "^6.3.1",
+    "angular-oauth2-oidc": "^4.0.2",
+    "angular-sortablejs": "^2.6.0",
+    "core-js": "^2.5.7",
     "file-saver": "^1.3.8",
-    "js-yaml": "^3.10.0",
-    "jsoneditor": "^5.13.1",
-    "lodash": "^4.17.4",
-    "ng-click-outside": "^3.2.0",
+    "js-yaml": "^3.12.0",
+    "jsoneditor": "^5.24.3",
+    "lodash": "^4.17.10",
+    "ng-click-outside": "^4.0.0",
     "ngx-clipboard": "^10.0.0",
-    "rxjs": "^5.5.2",
     "sortablejs": "^1.7.0",
-    "time-ago-pipe": "^1.2.1",
-    "tippy.js": "^2.4.1",
-    "webpack": "^3.8.1",
-    "zone.js": "^0.8.14"
+    "time-ago-pipe": "^1.3.2",
+    "tippy.js": "^2.5.4",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.5",
+    "@angular/compiler-cli": "^6.1.6",
+    "@angular/language-service": "^6.1.6",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
     "@types/node": "~6.0.60",
-    "ajv": "^6.0.0",
-    "codelyzer": "~4.1.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
+    "ajv": "^6.5.3",
+    "codelyzer": "~4.4.4",
+    "jasmine-core": "~3.2.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.0.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^2.0.2",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
     "karma-phantomjs-launcher": "^1.0.4",
-    "node-sass": "^4.8.3",
+    "node-sass": "^4.9.3",
     "protractor": "^5.4.0",
     "ts-node": "~3.2.0",
-    "tslint": "^5.7.0",
-    "tslint-angular": "^1.1.1",
-    "tslint-config-prettier": "^1.8.0",
-    "typescript": "~2.4.2"
+    "tslint": "^5.11.0",
+    "tslint-angular": "^1.1.2",
+    "tslint-config-prettier": "^1.15.0",
+    "typescript": "~2.9.2"
   },
   "prettier": {
     "singleQuote": true
