@@ -1,24 +1,23 @@
 import {
-    ChangeDetectorRef,
-    Component,
-    ComponentFactoryResolver,
-    Input,
-    OnChanges,
-    ReflectiveInjector,
-    SimpleChange,
-    Type,
-    ViewChild,
-    ViewContainerRef,
-    ViewRef
+  ChangeDetectorRef,
+  Component,
+  ComponentFactoryResolver,
+  Input,
+  OnChanges,
+  SimpleChange,
+  Type,
+  ViewChild,
+  ViewContainerRef,
+  ViewRef,
 } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/delay';
+import { Observable, of } from 'rxjs';
+import { delay } from 'rxjs/operators';
+import { Injector } from '@angular/core';
 
 @Component({
   selector: 'y-plain-list',
   templateUrl: './plain-list.component.html',
-  styleUrls: ['./plain-list.component.scss']
+  styleUrls: ['./plain-list.component.scss'],
 })
 export class PlainListComponent implements OnChanges {
   @Input() source: Observable<any[]>;
@@ -32,11 +31,15 @@ export class PlainListComponent implements OnChanges {
   data: any[];
   loaded = false;
   loading = false;
+  initialized = false;
   errorMessage = '';
 
   componentFactoryResolver: ComponentFactoryResolver;
 
-  constructor(componentFactoryResolver: ComponentFactoryResolver, private changeDetector: ChangeDetectorRef) {
+  constructor(
+    componentFactoryResolver: ComponentFactoryResolver,
+    private changeDetector: ChangeDetectorRef,
+  ) {
     this.componentFactoryResolver = componentFactoryResolver;
   }
 
@@ -50,11 +53,13 @@ export class PlainListComponent implements OnChanges {
 
   setDelayedLoading() {
     // Only set loading status if it takes longer than defined ms
-    Observable.of(0).delay(333).subscribe(() => {
-      if (!this.loaded) {
-        this.setLoading(true);
-      }
-    });
+    of(0)
+      .pipe(delay(this.initialized ? 750 : 350))
+      .subscribe(() => {
+        if (!this.loaded) {
+          this.setLoading(true);
+        }
+      });
   }
 
   setLoading(status: boolean): void {
@@ -63,6 +68,9 @@ export class PlainListComponent implements OnChanges {
 
   setLoaded(status: boolean): void {
     this.loaded = status;
+    if (status) {
+      this.initialized = true;
+    }
   }
 
   render(source: Observable<any[]>) {
@@ -78,15 +86,14 @@ export class PlainListComponent implements OnChanges {
         data.forEach(entry => {
           const inputProviders = [
             { provide: 'entry', useValue: entry },
-            { provide: 'entryEventHandler', useValue: this.entryEventHandler }
+            { provide: 'entryEventHandler', useValue: this.entryEventHandler },
           ];
-          const resolvedInputs = ReflectiveInjector.resolve(inputProviders);
-          const injector = ReflectiveInjector.fromResolvedProviders(
-            resolvedInputs,
-            this.entriesViewContainer.parentInjector
-          );
+          const injector = Injector.create({
+            providers: inputProviders,
+            parent: this.entriesViewContainer.parentInjector,
+          });
           const componentFactory = this.componentFactoryResolver.resolveComponentFactory(
-            this.entryRenderer
+            this.entryRenderer,
           );
           const entryTag = this.entryTagName
             ? document.createElement(this.entryTagName)
@@ -95,20 +102,20 @@ export class PlainListComponent implements OnChanges {
           this.entriesViewContainer.insert(component.hostView);
         });
 
-          if (!(this.changeDetector as ViewRef).destroyed) {
-              this.changeDetector.detectChanges();
-          }
+        if (!(this.changeDetector as ViewRef).destroyed) {
+          this.changeDetector.detectChanges();
+        }
       },
       error => {
         this.setLoaded(true);
         this.setLoading(false);
         this.errorMessage = error.message ? error.message : error.error;
         this.data = [];
-          if (!(this.changeDetector as ViewRef).destroyed) {
-              this.changeDetector.detectChanges();
-          }
+        if (!(this.changeDetector as ViewRef).destroyed) {
+          this.changeDetector.detectChanges();
+        }
         throw error;
-      }
+      },
     );
   }
 }
