@@ -2,41 +2,38 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ResourceQuotaEntryRendererComponent } from './resource-quota-entry-renderer.component';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ComponentCommunicationService } from '../../../../../../shared/services/component-communication.service';
-import { Subject } from 'rxjs/Subject';
-import { Observable } from 'rxjs/Observable';
+import { of, Subject } from 'rxjs';
 
 describe('ResourceQuotasEntryRendererComponent', () => {
   let component: ResourceQuotaEntryRendererComponent;
   let fixture: ComponentFixture<ResourceQuotaEntryRendererComponent>;
   let componentCommunicationService: ComponentCommunicationService;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        declarations: [ResourceQuotaEntryRendererComponent],
-        providers: [
-          ComponentCommunicationService,
-          {
-            provide: 'entry',
-            useValue: {
-              name: 'name',
-              pods: '3',
-              limits: {
-                memory: '1',
-                cpu: '1'
-              },
-              requests: {
-                memory: '1',
-                cpu: '1'
-              }
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ResourceQuotaEntryRendererComponent],
+      providers: [
+        ComponentCommunicationService,
+        {
+          provide: 'entry',
+          useValue: {
+            name: 'name',
+            pods: '3',
+            limits: {
+              memory: '1',
+              cpu: '1'
+            },
+            requests: {
+              memory: '1',
+              cpu: '1'
             }
-          },
-          { provide: 'entryEventHandler', useValue: {} }
-        ],
-        schemas: [CUSTOM_ELEMENTS_SCHEMA]
-      }).compileComponents();
-    })
-  );
+          }
+        },
+        { provide: 'entryEventHandler', useValue: {} }
+      ],
+      schemas: [CUSTOM_ELEMENTS_SCHEMA]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ResourceQuotaEntryRendererComponent);
@@ -57,7 +54,7 @@ describe('ResourceQuotasEntryRendererComponent', () => {
         disabled: true
       };
       spyOn(componentCommunicationService, 'observable$').and.returnValue(
-        Observable.of(subject.next(entry))
+        of(subject.next(entry))
       );
       fixture.whenStable().then(async () => {
         fixture.detectChanges();
@@ -76,7 +73,7 @@ describe('ResourceQuotasEntryRendererComponent', () => {
         disabled: false
       };
       spyOn(componentCommunicationService, 'observable$').and.returnValue(
-        Observable.of(subject.next(entry))
+        of(subject.next(entry))
       );
       fixture.whenStable().then(async () => {
         fixture.detectChanges();
@@ -95,7 +92,7 @@ describe('ResourceQuotasEntryRendererComponent', () => {
         disabled: true
       };
       spyOn(componentCommunicationService, 'observable$').and.returnValue(
-        Observable.of(subject.next(entry))
+        of(subject.next(entry))
       );
       fixture.whenStable().then(async () => {
         fixture.detectChanges();
@@ -114,7 +111,7 @@ describe('ResourceQuotasEntryRendererComponent', () => {
         disabled: true
       };
       spyOn(componentCommunicationService, 'observable$').and.returnValue(
-        Observable.of(subject.next(entry))
+        of(subject.next(entry))
       );
       fixture.whenStable().then(async () => {
         fixture.detectChanges();
