@@ -3,8 +3,7 @@ import { AppModule } from './../../../../../app.module';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { PodsEntryRendererComponent } from './pods-entry-renderer.component';
 import { ComponentCommunicationService } from '../../../../../shared/services/component-communication.service';
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
+import { of, Subject } from 'rxjs';
 
 describe('PodsEntryRendererComponent', () => {
   let component: PodsEntryRendererComponent;
@@ -124,7 +123,7 @@ describe('PodsEntryRendererComponent', () => {
       disabled: true
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
@@ -147,7 +146,7 @@ describe('PodsEntryRendererComponent', () => {
       disabled: true
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
@@ -170,7 +169,7 @@ describe('PodsEntryRendererComponent', () => {
       disabled: false
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
