@@ -3,40 +3,37 @@ import { AppModule } from './../../../../../app.module';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReplicaSetsEntryRendererComponent } from './replica-sets-entry-renderer.component';
 import { ComponentCommunicationService } from '../../../../../shared/services/component-communication.service';
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
+import { Subject, of } from 'rxjs';
 
 describe('ReplicaSetsEntryRendererComponent', () => {
   let component: ReplicaSetsEntryRendererComponent;
   let fixture: ComponentFixture<ReplicaSetsEntryRendererComponent>;
   let componentCommunicationService: ComponentCommunicationService;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        imports: [AppModule],
-        providers: [
-          [{ provide: APP_BASE_HREF, useValue: '/my/app' }],
-          [
-            {
-              provide: 'entry',
-              useValue: {
-                objectMeta: {
-                  name: 'name'
-                },
-                pods: {
-                  warnings: [],
-                  pending: []
-                }
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [AppModule],
+      providers: [
+        [{ provide: APP_BASE_HREF, useValue: '/my/app' }],
+        [
+          {
+            provide: 'entry',
+            useValue: {
+              objectMeta: {
+                name: 'name'
+              },
+              pods: {
+                warnings: [],
+                pending: []
               }
             }
-          ],
-          [{ provide: 'entryEventHandler', useValue: {} }],
-          ComponentCommunicationService
-        ]
-      }).compileComponents();
-    })
-  );
+          }
+        ],
+        [{ provide: 'entryEventHandler', useValue: {} }],
+        ComponentCommunicationService
+      ]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ReplicaSetsEntryRendererComponent);
@@ -59,7 +56,7 @@ describe('ReplicaSetsEntryRendererComponent', () => {
       disabled: true
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
@@ -82,7 +79,7 @@ describe('ReplicaSetsEntryRendererComponent', () => {
       disabled: true
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
@@ -105,7 +102,7 @@ describe('ReplicaSetsEntryRendererComponent', () => {
       disabled: false
     };
     spyOn(componentCommunicationService, 'observable$').and.returnValue(
-      Observable.of(subject.next(entry))
+      of(subject.next(entry))
     );
     expect(component.disabled).toEqual(false);
     fixture.whenStable().then(async () => {
