@@ -4,7 +4,8 @@ import { EnvironmentInfo } from '../environment-info';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 
 import * as _ from 'lodash';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class EnvironmentsService {
@@ -17,40 +18,44 @@ export class EnvironmentsService {
   public getEnvironments(): Observable<EnvironmentInfo[]> {
     return this.http
       .get<any>(AppConfig.k8sApiServerUrl + 'namespaces?labelSelector=env=true')
-      .map(
-        response => {
-          const environments: EnvironmentInfo[] = [];
+      .pipe(
+        map(
+          response => {
+            const environments: EnvironmentInfo[] = [];
 
-          if (response.items) {
-            response.items.forEach(namespace => {
-              if (namespace.status.phase === 'Active') {
-                environments.push(
-                  new EnvironmentInfo(
-                    namespace.metadata.uid,
-                    namespace.metadata.name
-                  )
-                );
-              }
-            });
-          }
-          return environments;
-        },
-        err => console.log(err)
+            if (response.items) {
+              response.items.forEach(namespace => {
+                if (namespace.status.phase === 'Active') {
+                  environments.push(
+                    new EnvironmentInfo(
+                      namespace.metadata.uid,
+                      namespace.metadata.name
+                    )
+                  );
+                }
+              });
+            }
+            return environments;
+          },
+          err => console.log(err)
+        )
       );
   }
 
   public getEnvironment(id: string): Observable<EnvironmentInfo> {
     return this.http
       .get<any>(AppConfig.k8sApiServerUrl + 'namespaces/' + id)
-      .map(response => {
-        if (!_.isEmpty(response.metadata)) {
-          return new EnvironmentInfo(
-            response.metadata.uid,
-            response.metadata.name
-          );
-        }
-        return response;
-      });
+      .pipe(
+        map(response => {
+          if (!_.isEmpty(response.metadata)) {
+            return new EnvironmentInfo(
+              response.metadata.uid,
+              response.metadata.name
+            );
+          }
+          return response;
+        })
+      );
   }
 
   public createEnvironment(environmentName: string) {
@@ -62,13 +67,15 @@ export class EnvironmentsService {
         .post<any>(`${AppConfig.k8sApiServerUrl}namespaces`, body, {
           headers: new HttpHeaders().set('Content-Type', 'application/json')
         })
-        .map(response => {
-          if (!_.isEmpty(response.metadata)) {
-            this.envChangeStateEmitter$.emit(true);
+        .pipe(
+          map(response => {
+            if (!_.isEmpty(response.metadata)) {
+              this.envChangeStateEmitter$.emit(true);
+              return response;
+            }
             return response;
-          }
-          return response;
-        });
+          })
+        );
     }
   }
 
@@ -76,10 +83,12 @@ export class EnvironmentsService {
     if (environmentName) {
       return this.http
         .delete(`${AppConfig.k8sApiServerUrl}namespaces/${environmentName}`)
-        .map(response => {
-          this.envChangeStateEmitter$.emit(true);
-          return response;
-        });
+        .pipe(
+          map(response => {
+            this.envChangeStateEmitter$.emit(true);
+            return response;
+          })
+        );
     }
   }
 }
