@@ -7,12 +7,11 @@ import { By } from '@angular/platform-browser';
 import { CreatePresetModalComponent } from './create-preset-modal.component';
 import { IdpPresetsService } from '../idp-presets.service';
 import { ComponentCommunicationService } from '../../../../shared/services/component-communication.service';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
+import { Observable, of } from 'rxjs';
 
 class IdpPresetsServiceMock {
   public createIdpPreset(data) {
-    return Observable.of({});
+    return of({});
   }
 }
 
@@ -23,19 +22,17 @@ describe('CreatePresetModalComponent', () => {
   let header: DebugElement;
   let buttonSaveClasses: any;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        imports: [FormsModule],
-        providers: [
-          { provide: APP_BASE_HREF, useValue: '/my/app' },
-          { provide: IdpPresetsService, useClass: IdpPresetsServiceMock },
-          ComponentCommunicationService
-        ],
-        declarations: [CreatePresetModalComponent]
-      }).compileComponents();
-    })
-  );
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [FormsModule],
+      providers: [
+        { provide: APP_BASE_HREF, useValue: '/my/app' },
+        { provide: IdpPresetsService, useClass: IdpPresetsServiceMock },
+        ComponentCommunicationService
+      ],
+      declarations: [CreatePresetModalComponent]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(CreatePresetModalComponent);
@@ -240,7 +237,7 @@ describe('CreatePresetModalComponent', () => {
     const spyCreate = spyOn(
       IdpPresetsServiceMockStub,
       'createIdpPreset'
-    ).and.returnValue(Observable.of({}));
+    ).and.returnValue(of({}));
 
     // when
     fixture.detectChanges();
