@@ -1,12 +1,13 @@
 import { List } from '../../shared/datamodel/k8s/generic/list';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { AppConfig } from '../../app.config';
 import {
   IMicroFrontend,
   MicroFrontend
 } from '../../shared/datamodel/k8s/microfrontend';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class ExtensionsService {
@@ -16,20 +17,24 @@ export class ExtensionsService {
     const resourceUrl = `${
       AppConfig.k8sApiServerUrl_ui
     }namespaces/${namespaceId}/microfrontends`;
-    return this.http.get<List<IMicroFrontend>>(resourceUrl).map(res => {
-      return res.items.map(item => {
-        return new MicroFrontend(item);
-      });
-    });
+    return this.http.get<List<IMicroFrontend>>(resourceUrl).pipe(
+      map(res => {
+        return res.items.map(item => {
+          return new MicroFrontend(item);
+        });
+      })
+    );
   }
 
   getClusterExtensions(): Observable<any> {
     const resourceUrl = `${AppConfig.k8sApiServerUrl_ui}clustermicrofrontends`;
-    return this.http.get<List<IMicroFrontend>>(resourceUrl).map(res => {
-      return res.items.map(item => {
-        return new MicroFrontend(item);
-      });
-    });
+    return this.http.get<List<IMicroFrontend>>(resourceUrl).pipe(
+      map(res => {
+        return res.items.map(item => {
+          return new MicroFrontend(item);
+        });
+      })
+    );
   }
 
   isUsingSecureProtocol(url: string) {
