@@ -1,13 +1,14 @@
 import { Component, Injector, Input, OnInit, OnDestroy } from '@angular/core';
 import { AbstractKubernetesEntryRendererComponent } from '../../abstract-kubernetes-entry-renderer.component';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 import { ComponentCommunicationService } from '../../../../../shared/services/component-communication.service';
 
 @Component({
   selector: 'app-secrets-entry-renderer',
   templateUrl: './secrets-entry-renderer.component.html'
 })
-export class SecretsEntryRendererComponent extends AbstractKubernetesEntryRendererComponent
+export class SecretsEntryRendererComponent
+  extends AbstractKubernetesEntryRendererComponent
   implements OnInit, OnDestroy {
   constructor(
     protected injector: Injector,
