@@ -13,7 +13,7 @@ import { SecretsEntryRendererComponent } from './secrets-entry-renderer/secrets-
 import { SecretsHeaderRendererComponent } from './secrets-header-renderer/secrets-header-renderer.component';
 import { ComponentCommunicationService } from '../../../../shared/services/component-communication.service';
 import { DataConverter } from '@kyma-project/y-generic-list';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-secrets',
