@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { AppConfig } from '../../../../app.config';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class K8sResourceEditorService {
@@ -9,13 +10,15 @@ export class K8sResourceEditorService {
   updateResource(jsonData) {
     return this.httpClient
       .put(`${AppConfig.k8sServerUrl}${jsonData.metadata.selfLink}`, jsonData)
-      .map(
-        data => {
-          return data;
-        },
-        err => {
-          throw err;
-        }
+      .pipe(
+        map(
+          data => {
+            return data;
+          },
+          err => {
+            throw err;
+          }
+        )
       );
   }
 }
